requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
aiofiles>=23.2.0
pytz>=2024.1
requests>=2.31.0
python-dotenv>=1.0.0
